refactor: 공용 컴포넌트 스타일로 수정(#8)
