+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import Layout from '@layout/Layout';
 import Home from '@pages/Home/Home';
-import ChatRoomCreatedPage from '@pages/chatRoomCreatedPage';
+import ChatRoomCreatedPage from '@pages/chatRommCreated/chatRoomCreatedPage';
 import GlobalStyles from '../src/styles/GlobalStyles';
 import ChatRoomExitPage from '@pages/chatRoomExit/chatRoomExitPage';
